Fully Functioning UI
The UI is mostly complete. All it needs is access to things like the player character, enemy spawning sequence, enemy behavior, and potentially the gamemode. For now the BGM has been reverted back to a single looping track as there doesn't seem to be enough time to fully implement smooth transitions.